// gravitee-apim-portal-webui/projects/portal-webclient-sdk/src/lib/model/alertStatusResponse.ts
/**
 * Gravitee.io Portal Rest API
 * API dedicated to the devportal part of Gravitee
 *
 * Contact: contact@graviteesource.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AlertStatusResponse { 
    /**
     * Number of available plugins.
     */
    available_plugins?: number;
    /**
     * Is alerting enabled
     */
    enabled?: boolean;
}