// gravitee-apim-portal-webui/projects/portal-webclient-sdk/src/lib/model/alertTimeUnit.ts
/**
 * Gravitee.io Portal Rest API
 * API dedicated to the devportal part of Gravitee
 *
 * The version of the OpenAPI document: 3.14.0-SNAPSHOT
 * Contact: contact@graviteesource.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type AlertTimeUnit = 'SECONDS' | 'MINUTES' | 'HOURS';

export const AlertTimeUnit = {
    SECONDS: 'SECONDS' as AlertTimeUnit,
    MINUTES: 'MINUTES' as AlertTimeUnit,
    HOURS: 'HOURS' as AlertTimeUnit
};