// gravitee-apim-portal-webui/projects/portal-webclient-sdk/src/lib/model/alertType.ts
/**
 * Gravitee.io Portal Rest API
 * API dedicated to the devportal part of Gravitee
 *
 * The version of the OpenAPI document: 3.14.0-SNAPSHOT
 * Contact: contact@graviteesource.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type AlertType = 'STATUS' | 'RESPONSE_TIME';

export const AlertType = {
    STATUS: 'STATUS' as AlertType,
    RESPONSETIME: 'RESPONSE_TIME' as AlertType
};